"loading screen"
Gave it some fight mechanics, and a funny loading screen. All fight mechanics are completely differnt for each class that is chosen.
made a maze for players to go through. the directions are left and right but what happens in that direction is RNG, maze not finished. Will finish tomorrow.